Fix all Clippy warnings to pass CI
- Add Default impls for all pipeline stages and PreviewSlot (new_without_default)
- Fix excessive float precision in fast_powf (8.2629582e-8 -> 8.262_958e-8)
- Replace % N == 0 with .is_multiple_of(N) throughout
- Add # Safety docs to unsafe extern C X-Plane entry points
- Fix overindented doc list continuation in headtrack-output-wine
- Remove redundant f32 casts in headtrack-output-simconnect
- Use slice iteration instead of index loops in tester and GUI
- Collapse nested if statements in headtrack-gui
- Use &Path instead of &PathBuf in headtrack-engine function signature
- Add StepFrame type alias for complex return type in tracker

+++ crates/headtrack-output-simconnect/src/lib.rs
@@ -63,12 +63,12 @@ use tracing::{info, warn};
 /// Signs and axis mapping are empirical — verify in-cockpit and adjust.
 fn pose_to_simconnect(pose: &Pose) -> [f32; 6] {
     [
-        pose.x as f32 / 1000.0,
-        pose.y as f32 / 1000.0,
-        pose.z as f32 / 1000.0,
-        pose.pitch as f32, // SimConnect uses degrees directly for these
-        pose.roll as f32,
-        pose.yaw as f32,
+        pose.x / 1000.0,
+        pose.y / 1000.0,
+        pose.z / 1000.0,
+        pose.pitch, // SimConnect uses degrees directly for these
+        pose.roll,
+        pose.yaw,
     ]
 }
